[ feat ] get func for api
